Add EXTRACT_BIT and restrict z to be an unsigned short

+++ programs/program1/src/satisfiability.c
@@ -1,12 +1,12 @@
 #include "satisfiability.h"
 
-bool circuit_one( uint32_t pid, uint32_t z )
+bool circuit_one( uint32_t pid, uint16_t z )
 {
     //! @todo Implement.
     return pid + z;
 }
 
-bool circuit_two( uint32_t pid, uint32_t z )
+bool circuit_two( uint32_t pid, uint16_t z )
 {
     //! @todo Implement.
     return pid + z;
